Moved test for ContainerInfo into right project
